requirements.txt
# Core Web App
fastapi==0.115.12
streamlit==1.38.0

# Google OAuth + Drive
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0

# LLM & Embeddings
openai==1.86.0
sentence-transformers==2.5.1

# Vector Search
faiss-cpu

# File Parsing
PyMuPDF==1.24.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.3.0

# Utilities
requests==2.32.4
python-dotenv==1.1.0
numpy==2.3.0
protobuf==5.26.1
uvicorn[standard]==0.29.0
